Extracted PlotOptimals.py supporting functions into ShapeFunctions.py for readability
