+++ .gitignore
@@ -1,0 +1,2 @@
+map.html
+data.geojson
